+++ src/main/kotlin/sparta/nbcamp/wachu/domain/common/QueryDslUtils.kt
@@ -1,0 +1,26 @@
+package sparta.nbcamp.wachu.domain.common
+
+import com.querydsl.core.types.Expression
+import com.querydsl.core.types.Order
+import com.querydsl.core.types.OrderSpecifier
+import com.querydsl.core.types.dsl.EntityPathBase
+import com.querydsl.core.types.dsl.PathBuilder
+import org.springframework.data.domain.Pageable
+
+object QueryDslUtils {
+
+    @Suppress("UNCHECKED_CAST")
+    fun getOrderSpecifier(
+        pageable: Pageable,
+        path: EntityPathBase<*>
+    ): Array<OrderSpecifier<*>> {
+        val pathBuilder = PathBuilder(path.type, path.metadata)
+
+        return pageable.sort.toList().map { order ->
+            OrderSpecifier(
+                if (order.isAscending) Order.ASC else Order.DESC,
+                pathBuilder.get(order.property) as Expression<Comparable<*>>
+            )
+        }.toTypedArray()
+    }
+}
